portal/pyproject.toml
[project]
name = "portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "kubernetes>=35.0.0",
    "python-keycloak>=7.0.3",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
